Ignore unreceived ack for commands
For some reason Telycam PTZ camera doesn't always reply with
ack for each command. This is true randomly after camera position
has been saved. To avoid total freeze of control, decrease the wait
time to get an ack to 1 second and if ack is not received, act as
if nothing has happened.